Address feedback comments
